requirements.txt
asgiref==3.7.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
crispy-bootstrap5==0.7
cryptography==39.0.1
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.52.0
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
oauthlib==3.2.2
packaging==23.0
pluggy==1.0.0
psycopg==3.1.8
psycopg-binary==3.1.8
pycparser==2.21
pygame==2.4.0
PyJWT==2.6.0
pytest==7.2.2
python3-openid==3.2.0
pytz==2023.3
requestes==0.0.1
requests==2.30.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
urllib3==2.0.2
whitenoise==6.4.0
